pyproject.toml
[tool.poetry]
name = "gs-hackathon"
version = "0.1.0"
description = ""
authors = ["Rafał Opiłowski <rafalopilowski1@outlook.com>", "Adam Dański", "Maciej Lewandowski"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.6"}
loguru = "^0.7.3"
openapi-core = "^0.19.4"
openapi3 = "^1.8.2"
requests = "^2.32.3"
jinja2 = "^3.1.4"
python-owasp-zap-v2-4 = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
